Trac #31619: add "classical" argument to matrix block_ldlt().
This commit adds a "classical" flag to block_ldlt() that tells it
to compute a standard non-block, non-pivoting LDL^{T} factorization.
Such a factorization may not exist, but you are welcome to try it.

The main reason for introducing this is that it almost entirely
replaces the indefinite_factorization() method using the existing code
for block_ldlt(). The one feature indefinite_factorization() has that
block_ldlt(classical=True) does NOT have is the ability to operate on
symmetric matrices that are not Hermitian. For Cholesky decompositions
and positive-(semi)definite testing this makes no difference, but it
is technically possible to decompose a symmetric (but not Hermitian)
matrix with complex entries. Likewise for symmetric matrices over some
finite fields where "conjugate" is either undefined or not what you'd
expect it to be. The indefinite_factorization() method may still be
useful in those cases.

In the applications where block_ldlt(classical=True) CAN be used, it
will be a bit nicer since we don't have to worry about what kind of
symmetry the matrix has. For example, you can't waste time running the
algorithm="hermitian" indefinite factorization on a symmetric matrix
whose algorithm="symmetric" result is already cached.