Trac #31619: support is_square() for constant non-relational expressions.
The is_square() method for ring elements is supposed to return True if
the given element is the square of another element in the same ring.
This of course all goes to hell in the symbolic ring; nevertheless,
we are currently returning things like

  sage: SR(5).is_square()
  False

even though

  sage: sqrt(5) in SR
  True

This commit updates is_square() to return True for all constant
non-relational expressions, since you can just call sqrt() on them to
find an element that, when squared, gives you what you started with.