Trac #31594: fix an interface test with newer versions of giac.
We have a test in interfaces/giac.py that passes a string to giac()
asking it to compute an indefinite integral. In newer versions of
giac, the format of that result has changed -- but it was always
complicated. Too complicated to check by differentiating, subtracting
the original expression, and comparing with zero, in fact.

Since the point of the test is to check that we can pass commands to
giac(), I see no reason to use such a complicated integral as a test
case. Instead I've changed it to use sin(x)^2 and cos(x)^2, whose
integrals sum to x and can be compared (as symbolic expressions) with
that expected result quickly.