Trac #31594: support giac-1.7.x in symbolic integration tests.
The output of a symbolic integral in symbolic/integration/integral.py
now depends on the version of giac installed; newer versions bring
improved simplifications. To allow the test case in question to pass
with all versions of giac, this commit checks for symbolic expression
equality between the actual and expected results rather than string
equality.