Trac #31594: simplify two libgiac simplify() examples.
There are two examples in libs/giac/giac.pyx that call
libgiac.simplify on exp(I*pi/5)^3. Older versions of giac return the
an abstract result containing "rootof", but newer versions return the
actual complex roots. To allow the test case to pass with all versions
of giac, and since the point of the example is simply to exercise the
GiacFunction class, this commit simplifies the call to,

  libgiac.simplify(exp(I*pi))

which both 1.6.x and 1.7.x versions of giac are able to recognize
explicitly as -1.

The test in question was duplicated in the docscring for the
GiacFunctionNoEV class, but I can discern no reason for it being
there. I've simply removed that instance.