Introduce a mechanism for creating temporary variables with unique names
In various places in sagemath, temporary symbolic variables are
needed. Previously, these variables were created in an ad-hoc manner
and could clash with user-defined variables. This commit introduces a
unified mechanism for creating these variables and cleaning up
afterwards (removing the variable from the list of symbols in the
symbolic ring and forgetting any assumptions about it).