Added integrate_vector_N method
A method was added to allow the user to use a Gauss-Legendre integrator while specifying the number
of nodes desired. This is in constrast to the existing method integrate_vector which finds an
appropriate number of nodes based on a desired error bound and a an error heuristic.
As per comments by nbruin on ticket #30698, it is desirable to have this as a separate method
rather than as an option within the existing method as the method is then set at call.
It is my opinion that the name integrate_vector_N is not the best, and so improving it as this
point in time would be useful.