Added _bounding_data_list method
As part of determining the number of nodes required to achieve a given error bound when applying Gauss-Legendre integration, a bound on the integrand is required over a certain region. As part of the
method designed to achieve this, we require information of the minimal polynomial of the integrand. This hidden method provides that information.