Trac #32037: update the doctest for the DirichletGroup(5)[1] L-function.
Due to a change in libLfunction's API, this L-function is now
automatically determined to be complex, which means that we search for
zeros below the real axis as well. As a result, the first zero (in
magnitude) is now negative. (In the past, it was ignored without
do_negative=True). We update the doctest to include it.