Trac #32188: test rank() over GF(2^4)[a] against an expected value.
The existing random test,

  sage: while A.rank() != 1000:
  ....:     A = random_matrix(K, 1000, 1000)

tests that we can generate a full-rank matrix, but doesn't really
check anything else; the test would also pass if you replaced A.rank()
by ZZ.random_element(). To keep the test efficient while still testing
a known quantity, this commit replaces the test with a call to rank()
on a random 10-by-10 unimodular matrix that is supposed to have rank
ten.