Trac #32234: use domain=float for fast callables used by numpy.
If we pass a fast_callable() to numpy without specifying its domain,
we run the risk of hitting a NaN and triggering an error like,

  TypeError: ufunc 'isnan' not supported for the input types, and the
  inputs could not be safely coerced to any supported types according
  to the casting rule ''safe''

Our numeric routines (in particular, find_root) are not expecting
this. Instead, they want numpy to detect the NaN and give up,
returning a tuple that contains some information about where we
stopped. To let numpy detect the NaN, it seems to suffice to pass
domain=float to the fast_callable() function. We do this for both
the symbolic find_local_minimum() and find_root().