Trac #8450: use fast_callable() in setup_for_eval_on_grid().
By replacing the old fast_float() with fast_callable(), we free
ourselves from the constraints of domain=float and can now plot
functions involving complex numbers. This makes use of the new
FastCallablePlotWrapper class to perform computations over CDF
with the results being forced to "float" in a manner that is
suitable for plotting.