Addition of abel_jacobi methods and edits to architecture.
This commit contains a large amount of work, combined together.

1) The structure of many method that use upstairs_edge have been changed to allow for the
flexibilty that now an edge need not correspond to an edge of the voronoi diagram, but can instead
be specified by a tuple (z_start, z_end). This mean that methods such as
simple_vector_line_integral can now integrate along paths that are not part of the skeleton.

2) Methods for the calculation of the Abel-Jacobi map, namely _integrate_differentials_iteratively,
_aj_based, abel_jacobi and reparameterise_differential_minpoly, have been added.

3) __init__ has been changed to now include the branchpoints of the output of cohomology_basis()
in self.branch_locus. The branch locus of just self.f is now stored in self._f_branch locus. This
is useful when calculating the Abel-Jacobi map. As a result the voronoi diagram for a given curve
may now be different, and this is reflected in the examples and tests for certain methods.

4) A method for the upstairs graph has been added, which is useful when calculating the Abel-Jacobi
map.

5) Matrix of integral values now writes the output of line integrals to self._integral_dict if the
differentials are the cohomology basis. This saves on recalculation when calling the Abel-Jacobi
map, which can be useful when calculating these integrals is slow.

6) Methods to help with the manipulation of divisors have been added. Namely the methods
places_at_branch_locus, strong_approximation, and divisor_to_divisor_list.

7) A method reduce_over_period_lattice has been added to help with the interpretation of results
from abel_jacobi, as these will be path dependent.