Trac #31275: use boolean SINGULAR variable.
In preparation for a system copy of Singular, this commit adds a new
patch that makes the upstream "SINGULAR" variable (defined in
qepcadrc) boolean. Previously, it was either a path, or disabled (when
no path was given). The path however was only used to execute
"Singular", which can safely be assumed to live in the user's $PATH
these days -- especially within SageMath.

The new patch makes "SINGULAR yes" run "Singular" from the $PATH, and
we update spkg-install.in to make use of it.