Trac #29024: always obtain the absolute path to libSingular.
Trusting dlopen() to look up its argument in the system's library
search path did not work as well as expected. At least on macOS --
and probably elsewhere -- the autoconf test succeeds but the real
call in singular.pyx fails, requiring us to pass it the absolute
path instead. That path can be obtained from pkg-config.

We're already using pkg-config for a version check, and the name-only
behavior of dlopen() is not standard, so relying on pkg-config to get
singular's libdir does not seem so bad in this case. To simplify the
implementation, this commit uses the libdir obtained from pkg-config
unconditionally, and not just on macOS. It also adds some (untested)
code in preparation for a system copy of Singular on Cygwin.