Trac #32234: use domain=float in a few more fast_callables.
In an earlier commit, some numerical routines were converted from the
old fast_float() to fast_callable(). Since these interface with
external libraries that expect floats, there is no harm in passing
domain=float when we construct the fast callable. This should provide
a minor speedup; and is probably closer to the old behavior anyway.