Trac #32656: don't list psutil as a dependency of pyrsistent.
Perusing the pyrsistent source code, we see that the psutil package is
needed only to run the pyrsistent test suite -- not in the course of
normal usage. Here we remove "psutil" from its list of dependencies,
pending the removal of psutil itself, so that `git grep psutil` does
not return false positives.