Trac #8450: use CDF for symbolic expression fast-callables.
With the introduction of the FastCallablePlotWrapper, we "expect" to
produce both complex intermediate results and complex noise. A few
other cases are handled lower in the plotting machinery, but symbolic
expressions are converted to fast-callables early, via their
_plot_fast_callable() method. For consistency with what
setup_for_eval_on_grid() now does, we make _plot_fast_callable()
return a fast-callable over domain=CDF.

The doctest for trac 19797 was removed because its purpose was to
ensure that some non-real terms don't arise during plotting. Now
that plots are over CDF, the test is unhelpful. To return a real
result, a FastCallablePlotWrapper around the the CDF fast-callable
should be used.