Updates and bug fixes.
The divisor_to_divisor_list method was made more general, and the reduce_over_period_lattice was made to allow for the period matrix to be normalised (helpful if calculating the RCV).
The _aj_based method was updated to deal with errors caused when the point take the map to was already a vertex.