Trac #32841: drop zn_poly interval_products() implementation.
Our interval_products_wrapper() wrapped three different implementations:

  * ntl_interval_products (ZZ_p version)
  * ntl_interval_products (zz_p version)
  * zn_poly_interval_products

But "the zn_poly version occasionally fails," and when that happens,
the NTL implementation is used anyway. Moreover the wrapper takes a
"force_ntl" parameter that allows the user to skip the zn_poly attempt
entirely.

This is the only place that zn_poly is actually used in sagelib
anymore, so this commit removes the underlying zn_poly implementation,
and eliminates the "force_ntl" parameter that was propagated
throughout the "hypellfrob" code. Now NTL is always used, as if
force_ntl was enabled.