Trac #32511: avoid passing Expression objects to minimize_constrained().
Despite its documentation (which says that it accepts only symbolic or
python functions), our EXAMPLES encourage users to pass in symbolic
expressions to minimize_constrained() as both the objective function
and the constraints.

In the future, when the arguments() method is deprecated and
subsequently removed from non-callable expressions, these examples
will become errors. In anticipation, we fix the examples now to follow
the documentation and pass in callable symbolic functions.

This is less of an issue for the constraints than it is for the
objective function, since the constraints implicitly use the same
arguments as the objective function, but for maximum clarity it's best
to document the use of symbolic functions for the contraints as well.