Trac #32964: fix failing doctest in functions.orthogonal_polys.
One of the examples in this module is expecting a doctest warning, but
that warning is only emitted the first time it happens. When testing
with

  sage -t --file-iterations=2

the warning does not appear in the second pass, and the test (that is
expecting it) fails. This commit adds some more "..." to support both
possible outcomes.