Trac #33024: fix sparse Cholesky when cvxopt is not available.
When --disable-cvxopt is used, we cannot use its cholmod()
implementation of sparse Cholesky. I tried to handle this case the
first time around, but it looks like I installed cvxopt to test out
the common case (when cvxopt is present) and then never went back to
double-check everything at the end. The fix is trivial.