Trac #32891: fix a failing hyperbolic model doctest.
One of the tests in sage.geometry.hyperbolic_space.hyperbolic_model
is testing the wrong property (or the right property of the wrong
model). In the Poincare Disk (PD) model, we should be testing that
the endpoints of a geodesic lie in the unit disk, rather than in
the upper half-plane.