Trac #32890: fix a failing random doctest in sage.stats.hmm.chmm.
One of the doctests in this module is checking that the number of
iterations taken in the baum_welch() method is between 4 and 199,
inclusive. However, with one particular random seed, it only takes 3
iterations -- and could potentially take as many as 500 (the method's
max_iter value). Here we update the check to ensure that the number
of iterations is between 1 and 500, inclusive.