Trac #30875: randomize random elements from AA.
Previously there was no random_element() method for the field of
algebraic real numbers, and as a result, AA.random_element() would
return only integers between -2 and 2, inclusive. Here we add such a
method that simply delegates to QQbar.random_element(), and takes the
real part of the result. I have avoided moving the method into the
superclass AlgebraicField_common because that approach does not seem
to fit the design of this module, even though it might ultimately have
been cleaner.