Trac #8450: replace a flaky test.
One test in sage.plot.plot is looking for a glob of floating point
numbers as the result of generating some plot points for sin(x) with
randomize=False. With this test now failing on OSX, it's a good time
to re-evaluate whether or not such a test is actually useful. This
commit changes the test to confirm that the x-delta between the points
does in fact remain consistent; that is, that randomize=False does
what it's supposed to. We're still looking at four floating point
numbers as the expected output, but they're all the same, and that's
something that a reader can actually verify.