Trac #33176: fix Cython warning in sage.matrix.matrix_modn_dense_template.
The linbox_det() method in this file creates a float/double variable
"d" and passes it by reference to the FFLAS/FFPACK pDet() method,
which in turn passes it by reference to the Det() method to be
overwritten with the return value from that function. This is all
fine, but Cython doesn't know what the uninitialized variable is
used for and throws warnings like,

  warning: sage/matrix/matrix_modn_dense_template.pxi:283:19: local
  variable 'd' referenced before assignment

Here we initialize it to zero to avoid the warnings.