Trac #33186: fix some Cython warnings in padics.
A pattern in our padics is to define two new "long" variables and then
initialize them with a call to ZZ_pX_min_val_coeff(). For example,

  cdef long val, index
  ZZ_pX_min_val_coeff(val, index, ...)

The variables are passed by reference and promptly overwritten, so
this is harmless, but Cython warns,

  warning: sage/rings/padics/padic_ZZ_pX_CA_element.pyx:759:28: local
  variable 'val' referenced before assignment
  warning: sage/rings/padics/padic_ZZ_pX_CA_element.pyx:759:33: local
  variable 'index' referenced before assignment

To avoid the warnings, we initialize these variables to zero.