Trac #33201: fix Cython warnings for dense float/double matrices.
In our template for dense float/double matrices mod-n, we define a few
variables "zero", "one", and "d" that are passed by reference to C
functions to be overwritten. Cython however is not smart enough to
know that these uninitialized variables are about to be initialized in
C, and throws warnings like

  warning: sage/matrix/matrix_modn_dense_template.pxi:283:19: local
  variable 'd' referenced before assignment

To avoid the warnings, we initialize all of these variables (whose
type is either float or double, depending on the context) to zero.